Noch nicht fertig: Anpassungen CalendarDate

+++ WebInterface_210_0/MacroNamePanelController.m
@@ -15,7 +15,7 @@
 {
     MacroNamePanelController* 	theController;
     NSString* 			result = nil;
-    int				returnCode;
+    long				returnCode;
 
     theController = [[MacroNamePanelController alloc] init];
     [theController showWindow:nil];
